start implementing of complex alias definitions for reporting
