+++ app/src/Controller/ReportsController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\ReportDefinition;
 use App\Entity\User;
+use App\Service\ReportDataService;
 use App\Service\ReportFieldAliasService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -16,8 +17,9 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 class ReportsController extends AbstractController
 {
     #[Route('/reports/preview', name: 'app_report_preview', methods: ['POST'])]
-    public function preview(Request $request, EntityManagerInterface $em): Response
+    public function preview(Request $request, EntityManagerInterface $em, ReportDataService $reportDataService): Response
     {
+        /** @var User $user */
         $user = $this->getUser();
         $fieldAliases = ReportFieldAliasService::fieldAliases();
         $data = $request->request->all();
@@ -28,8 +30,8 @@ class ReportsController extends AbstractController
             'diagramType' => $data['config']['diagramType'] ?? 'bar',
             'xField' => $data['config']['xField'] ?? 'player',
             'yField' => $data['config']['yField'] ?? 'goals',
-            // Accept groupBy as array, fallback to empty array
             'groupBy' => $data['config']['groupBy'] ?? [],
+            'filters' => $data['config']['filters'] ?? [],
         ];
         if (!isset($fieldAliases[$config['xField']])) {
             $config['xField'] = array_key_first($fieldAliases);
@@ -37,7 +39,6 @@ class ReportsController extends AbstractController
         if (!isset($fieldAliases[$config['yField']])) {
             $config['yField'] = array_key_first($fieldAliases);
         }
-        // groupBy: ensure array
         if (!is_array($config['groupBy'])) {
             $config['groupBy'] = $config['groupBy'] ? [$config['groupBy']] : [];
         }
@@ -45,95 +46,22 @@ class ReportsController extends AbstractController
         $report->setIsTemplate(false);
         $report->setUser($user);
 
-        // Aggregation logic with multi-grouping support
-        $xField = $config['xField'] ?? 'player';
-        $yField = $config['yField'] ?? 'goals';
-        $diagramType = $config['diagramType'] ?? 'bar';
-        $groupBy = $config['groupBy'] ?? [];
-        $filters = $config['filters'] ?? [];
-
-        $qb = $em->getRepository(\App\Entity\GameEvent::class)->createQueryBuilder('e');
-        if (isset($filters['team'])) {
-            $qb->andWhere('e.team = :team')->setParameter('team', $filters['team']);
-        }
-        if (isset($filters['eventType'])) {
-            $qb->andWhere('e.gameEventType = :eventType')->setParameter('eventType', $filters['eventType']);
-        }
-        if (isset($filters['dateFrom'])) {
-            $qb->andWhere('e.timestamp >= :dateFrom')->setParameter('dateFrom', new \DateTimeImmutable($filters['dateFrom']));
-        }
-        if (isset($filters['dateTo'])) {
-            $qb->andWhere('e.timestamp <= :dateTo')->setParameter('dateTo', new \DateTimeImmutable($filters['dateTo']));
-        }
-        $events = $qb->getQuery()->getResult();
-
-        // Flexible Aggregation für beliebige X/Y (auch kategorisch)
-        $xValues = [];
-        $yValues = [];
-        $matrix = [];
-        $xyDebug = [];
-        foreach ($events as $event) {
-            $x = $this->retrieveFieldValue($event, $xField);
-            $y = $this->retrieveFieldValue($event, $yField);
-            $xyDebug[] = ['x' => $x, 'y' => $y];
-            $xValues[$x] = true;
-            $yValues[$y] = true;
-            if (!isset($matrix[$x][$y])) {
-                $matrix[$x][$y] = 0;
-            }
-            $matrix[$x][$y] += 1;
-        }
-        $xLabels = array_keys($xValues);
-        $yLabels = array_keys($yValues);
-        sort($xLabels);
-        sort($yLabels);
-
-        // IMMER vollständige Matrix: Labels = X, Datasets = Y
-        $labels = $xLabels;
-        $datasets = [];
-        foreach ($yLabels as $yVal) {
-            $data = [];
-            foreach ($xLabels as $xVal) {
-                $data[] = $matrix[$xVal][$yVal] ?? 0;
-            }
-            $datasets[] = [
-                'label' => (string)$yVal,
-                'data' => $data,
-            ];
-        }
+        $reportData = $reportDataService->generateReportData($config);
         $previewHtml = $this->renderView('report/_preview_chart.html.twig', [
             'report' => $report,
-            'labels' => $labels,
-            'datasets' => $datasets,
-            'diagramType' => $diagramType,
+            'labels' => $reportData['labels'],
+            'datasets' => $reportData['datasets'],
+            'diagramType' => $config['diagramType'],
         ]);
-        // Debug-Ausgabe als Teil der Response
+
         return $this->json([
             'preview' => $previewHtml,
-            'debug_labels' => $labels,
-            'debug_datasets' => $datasets,
-            'debug_xy' => $xyDebug,
+            'debug_labels' => $reportData['debug_labels'],
+            'debug_datasets' => $reportData['debug_datasets'],
+            'debug_xy' => $reportData['debug_xy'],
         ]);
     }
 
-    // Hilfsmethode für Vorschau wie in Api/ReportController
-    private function retrieveFieldValue($event, string $field): mixed
-    {
-        $fieldAliases = \App\Service\ReportFieldAliasService::fieldAliases();
-        if (isset($fieldAliases[$field]['value']) && is_callable($fieldAliases[$field]['value'])) {
-            return $fieldAliases[$field]['value']($event);
-        }
-        $getter = 'get' . ucfirst($field);
-        if (method_exists($event, $getter)) {
-            $value = $event->$getter();
-            if (is_object($value) && method_exists($value, '__toString')) {
-                return (string) $value;
-            }
-            return $value;
-        }
-        return null;
-    }
-    
     #[Route('/reports', name: 'app_report_list', methods: ['GET'])]
     public function list(EntityManagerInterface $em): Response
     {
@@ -160,7 +88,7 @@ class ReportsController extends AbstractController
         if ($isEdit && (!$report || ($report->getUser() && $report->getUser() !== $user && !$this->isGranted('ROLE_ADMIN')))) {
             throw $this->createAccessDeniedException();
         }
-        
+
         if ($request->isMethod('POST')) {
             $data = $request->request->all();
             $report->setName($data['name'] ?? '');
