+++ app/templates/report/_preview_chart.html.twig
@@ -6,11 +6,27 @@
     <script>
     (function() {
         let labels = {{ labels|json_encode|raw }};
+        const palette = [
+            'rgba(54, 162, 235, 0.5)',
+            'rgba(255, 99, 132, 0.5)',
+            'rgba(255, 206, 86, 0.5)',
+            'rgba(75, 192, 192, 0.5)',
+            'rgba(153, 102, 255, 0.5)',
+            'rgba(255, 159, 64, 0.5)'
+        ];
+        const borderPalette = [
+            'rgba(54, 162, 235, 1)',
+            'rgba(255, 99, 132, 1)',
+            'rgba(255, 206, 86, 1)',
+            'rgba(75, 192, 192, 1)',
+            'rgba(153, 102, 255, 1)',
+            'rgba(255, 159, 64, 1)'
+        ];
         let datasets = (
             {{ datasets|json_encode|raw }}
         ).map((ds, i) => Object.assign({
-            backgroundColor: 'rgba(54, 162, 235, 0.5)',
-            borderColor: 'rgba(54, 162, 235, 1)',
+            backgroundColor: palette[i % palette.length],
+            borderColor: borderPalette[i % borderPalette.length],
             borderWidth: 1
         }, ds));
         let swapped = false;
